AbstractMovableEntity.java: fix setters for direction. translate/moveTo now use double. Entity.java: position now use Point2D and Rectangle2D. Use double instead of integer. glutton.java: use double + fix formulas. Movable.java: use double

+++ src/engine/Movable.java
@@ -14,8 +14,8 @@ public interface Movable {
      */
     public void move(Direction d);
 
-    public void translate(int dx, int dy);
+    public void translate(double dx, double dy);
 
-    public void moveTo(int x, int y);
+    public void moveTo(double x, double y);
 
 }
